+++ lib/ui/workout/view_model/simple_view_model.dart
@@ -34,7 +34,7 @@ class SimpleViewModel extends BaseViewModel {
     final workout = Provider.of<WorkoutViewModel>(context, listen: false).currentWorkout;
     final now = DateTime.now().toIso8601String();
     try {
-      super.addOrUpdateSetResult(context, workout.items.first.id, workout.items.first.itemId, int.parse(repeatsController.text), double.parse(weightController.text).toInt(), 0, now);
+      super.addOrUpdateSetResult(context, workout.items.first.id, workout.items.first.itemId, int.parse(repeatsController.text), double.parse(weightController.text), 0, now);
     } on Exception {}
     _resetText();
   }
